test(folder-predictor): split LcppnFolderPredictor tests under 500-line cap (#177)
Remediation cycle 2 for the cycle-1 exit-reaudit FAIL (AC20). The cycle-1
coverage work pushed LcppnFolderPredictor_Tests.cs to 554 lines, over the
500-line file cap. Split the classify/beam-descent/abstention cases into a
new sibling LcppnFolderPredictor_Classify_Tests.cs, leaving
config/validation/training/build in the original file.

Result: LcppnFolderPredictor_Tests.cs 316 lines, LcppnFolderPredictor_Classify_Tests.cs
287 lines (both <= 500). All 21 cases preserved; LcppnFolderPredictor strict
coverage unchanged at 97.71%. Test-only change; no production file touched;
ManagerAsyncLazy.cs and the out-of-scope subsystems untouched. Toolchain green.
