+++ docs/features/active/2026-06-08-hierarchical-lcppn-folder-prediction-177/evidence/baseline/analyzers-baseline.2026-06-12T16-45.md
@@ -1,0 +1,12 @@
+# Baseline Analyzer Build (Cycle 2)
+
+Timestamp: 2026-06-12T16:57Z
+
+Command: msbuild TaskMaster.sln /t:Build /p:Configuration=Debug /p:Platform="Any CPU" /p:EnableNETAnalyzers=true /p:EnforceCodeStyleInBuild=true
+(Executed with VS18 Community MSBuild: "C:\Program Files\Microsoft Visual Studio\18\Community\MSBuild\Current\Bin\MSBuild.exe")
+
+EXIT_CODE: 0
+
+Output Summary:
+Build succeeded. 0 Warning(s), 0 Error(s). All projects (including UtilitiesCS.Test)
+built clean under the analyzer configuration. Baseline analyzer state is clean.
