+++ docs/features/active/2026-06-08-hierarchical-lcppn-folder-prediction-177/evidence/baseline/csharpier-baseline.2026-06-12T16-45.md
@@ -1,0 +1,11 @@
+# Baseline CSharpier Formatting (Cycle 2)
+
+Timestamp: 2026-06-12T16:55Z
+
+Command: dotnet tool run csharpier check .
+(CSharpier v1.2.6 uses the `check` subcommand; the plan-cited `--check` is the v0 form.)
+
+EXIT_CODE: 0
+
+Output Summary:
+Checked 1076 files in 3124ms. No unformatted files. Formatting baseline is clean.
