+++ docs/features/active/2026-06-08-hierarchical-lcppn-folder-prediction-177/evidence/baseline/linecount-baseline.2026-06-12T16-45.md
@@ -1,0 +1,11 @@
+# Baseline Line Count (Cycle 2)
+
+Timestamp: 2026-06-12T16:54Z
+
+Command: wc -l "UtilitiesCS.Test/EmailIntelligence/Bayesian/LcppnFolderPredictor_Tests.cs"
+
+EXIT_CODE: 0
+
+Output Summary:
+LcppnFolderPredictor_Tests.cs = 554 lines (over the 500-line cap; AC20 FAIL / finding F3).
+This is the single in-scope finding for cycle 2.
