+++ docs/features/active/2026-06-08-hierarchical-lcppn-folder-prediction-177/evidence/baseline/nullable-baseline.2026-06-12T16-45.md
@@ -1,0 +1,15 @@
+# Baseline Nullable / Type-Check Build (Cycle 2)
+
+Timestamp: 2026-06-12T16:58Z
+
+Command: msbuild TaskMaster.sln /t:Build /p:Configuration=Debug /p:Platform="Any CPU" /p:Nullable=enable /p:TreatWarningsAsErrors=true
+(Executed with VS18 Community MSBuild.)
+
+EXIT_CODE: 0
+
+Output Summary:
+Build succeeded. 0 Warning(s), 0 Error(s). CoreCompile was skipped as up-to-date
+(incremental), so the build is green under the nullable gate at baseline. Per the
+incremental nullable gate convention, any pre-existing unrelated CS8625 in other test
+files are noted as out-of-scope incremental exclusions, not failures introduced by this
+work; the current solution builds clean with TreatWarningsAsErrors=true.
