+++ docs/features/active/2026-06-08-hierarchical-lcppn-folder-prediction-177/evidence/qa-gates/containment.2026-06-12T16-45.md
@@ -1,0 +1,37 @@
+# Containment Verification (Cycle 2)
+
+Timestamp: 2026-06-12T17:13Z
+
+Command:
+- git diff --stat HEAD -- <each protected file>
+- git status --short -- '*.cs' (filtered)
+
+EXIT_CODE: 0
+
+Output Summary:
+
+## Protected files — zero diff (working tree vs HEAD)
+Each of the following showed an empty `git diff --stat` (no changes):
+- UtilitiesCS/EmailIntelligence/ClassifierGroups/ManagerAsyncLazy.cs (and the shared
+  `Manager` value type it defines — no separate Manager.cs file exists, and ManagerAsyncLazy.cs
+  is byte-identical)
+- UtilitiesCS/EmailIntelligence/ClassifierGroups/Triage/Triage.cs
+- UtilitiesCS/EmailIntelligence/ClassifierGroups/SpamBayes/SpamBayes.cs
+  (and UtilitiesCS/EmailIntelligence/Bayesian/SpamBayes.cs)
+- UtilitiesCS/EmailIntelligence/ClassifierGroups/Categories/CategoryClassifierGroup.cs
+- UtilitiesCS/EmailIntelligence/ClassifierGroups/MulticlassEngine.cs
+
+## Source files changed this cycle (only test files + test .csproj)
+- M  UtilitiesCS.Test/EmailIntelligence/Bayesian/LcppnFolderPredictor_Tests.cs (trimmed)
+- ?? UtilitiesCS.Test/EmailIntelligence/Bayesian/LcppnFolderPredictor_Classify_Tests.cs (new)
+- M  UtilitiesCS.Test/UtilitiesCS.Test.csproj (one <Compile Include> added)
+
+No production .cs file was modified (the `git status --short -- '*.cs'` filter excluding the
+two split test files returned no results). The remaining working-tree changes are
+documentation/evidence/plan/agent-memory artifacts (development-log.md, the cycle-2 plan,
+baseline and qa-gate evidence, and pre-existing upstream agent-memory edits made by the
+planner/reviewer, not by this executor and not production code).
+
+Containment invariant holds: zero diff to ManagerAsyncLazy.cs, the shared Manager value
+type, Triage.cs, SpamBayes.cs, CategoryClassifierGroup.cs, and MulticlassEngine.cs; the only
+changed source artifacts are the two LcppnFolderPredictor test files and UtilitiesCS.Test.csproj.
