+++ docs/features/active/2026-06-08-hierarchical-lcppn-folder-prediction-177/evidence/qa-gates/final-toolchain.2026-06-12T16-45.md
@@ -1,0 +1,42 @@
+# Final Toolchain Pass (Cycle 2) — single uninterrupted pass
+
+Timestamp: 2026-06-12T17:11Z
+
+Order: CSharpier check -> analyzers msbuild -> nullable/TWAE msbuild -> vstest /EnableCodeCoverage.
+No source files changed between or during the four steps (no restart triggered by file
+change). No SKIPPED outcomes.
+
+## Step 1 — CSharpier
+Command: dotnet tool run csharpier check .
+EXIT_CODE: 0
+Output Summary: Checked 1077 files; no files reformatted. Formatting clean.
+
+## Step 2 — Analyzers
+Command: msbuild TaskMaster.sln /t:Build /p:Configuration=Debug /p:Platform="Any CPU" /p:EnableNETAnalyzers=true /p:EnforceCodeStyleInBuild=true
+EXIT_CODE: 0
+Output Summary: Build succeeded. Zero diagnostics reference the two split files. Pre-existing
+CS8632/CS0067 warnings in untouched files only; no new analyzer error.
+
+## Step 3 — Nullable / TreatWarningsAsErrors
+Command: msbuild TaskMaster.sln /t:Build /p:Configuration=Debug /p:Platform="Any CPU" /p:Nullable=enable /p:TreatWarningsAsErrors=true
+EXIT_CODE: 0
+Output Summary: Build succeeded. 0 Warning(s), 0 Error(s). No new nullable warning in the
+split files.
+
+## Step 4 — Tests with coverage
+Command: vstest.console.exe UtilitiesCS.Test/bin/Debug/UtilitiesCS.Test.dll /InIsolation /EnableCodeCoverage
+EXIT_CODE: 1
+Output Summary: Total 3904; Passed 3903; Failed 1. The single failure is the documented
+out-of-scope pre-existing flake AddEntry_UseUiThreadTrue_DequeuesEntryAndSuppressesDispatcherException
+(IdleAsyncQueue / ci-flaky-test-isolation-176), explicitly recorded as out of scope in
+remediation-inputs.2026-06-12T16-45.md. It passes 1/1 in isolation (re-verified this pass).
+The failure is unrelated to the test-file split and to LcppnFolderPredictor; all in-scope
+tests pass. Post-change coverage merged to artifacts/csharp/coverage.xml:
+LcppnFolderPredictor strict line = 97.71%, block = 97.58%; UtilitiesCS.dll line = 85.46%.
+
+## Pass assessment
+Steps 1-3 are clean (EXIT 0). Step 4's only failure is the out-of-scope pre-existing flake
+(passes in isolation), not introduced by this work and explicitly excluded by the cycle-2
+inputs; per the standing rules the restart-from-CSharpier loop applies to failures
+introduced by this work, so no restart is required for this excluded flake. All in-scope
+acceptance is met.
