+++ docs/features/active/2026-06-08-hierarchical-lcppn-folder-prediction-177/evidence/qa-gates/p1-analyzers.2026-06-12T16-45.md
@@ -1,0 +1,23 @@
+# Phase 1 Analyzer Build (Cycle 2)
+
+Timestamp: 2026-06-12T17:05Z
+
+Command: msbuild TaskMaster.sln /t:Build /p:Configuration=Debug /p:Platform="Any CPU" /p:EnableNETAnalyzers=true /p:EnforceCodeStyleInBuild=true
+(VS18 Community MSBuild.)
+
+EXIT_CODE: 0
+
+Output Summary:
+Build succeeded. The build recompiled UtilitiesCS.Test (the two split files changed),
+surfacing 27 pre-existing warnings that were present before this work but masked in the
+incremental baseline build (CoreCompile skipped at baseline):
+- 24x CS8632 ("nullable annotation outside #nullable context") in untouched files:
+  ProgressTracker_Tests.cs, OlTableExtensions_Tests.cs, ConversationHelper_ExtendedTests.cs,
+  ManualFireTimerWrapper.cs.
+- 3x CS0067 ("event never used") in untouched files: SmartSerializable_Tests.cs,
+  StoreWrapperControllerTests.cs, SmartSerializableBase_Tests.cs.
+Zero diagnostics reference the two split files
+(LcppnFolderPredictor_Tests.cs / LcppnFolderPredictor_Classify_Tests.cs). These warnings
+are pre-existing in other files, out of scope per the incremental gate convention, and
+are not errors — the build succeeded with EXIT_CODE 0. No new analyzer error introduced
+by this work.
