+++ docs/features/active/2026-06-08-hierarchical-lcppn-folder-prediction-177/evidence/qa-gates/p1-csharpier.2026-06-12T16-45.md
@@ -1,0 +1,13 @@
+# Phase 1 CSharpier Formatting (Cycle 2)
+
+Timestamp: 2026-06-12T17:03Z
+
+Command: dotnet tool run csharpier format . (then `csharpier check .` to confirm stability)
+
+EXIT_CODE: 0
+
+Output Summary:
+`format .` formatted 1077 files; the subsequent `check .` reported all 1077 files
+already formatted (no files reformatted on the final pass). Only the two intended test
+files are modified/added (LcppnFolderPredictor_Tests.cs modified,
+LcppnFolderPredictor_Classify_Tests.cs new). Formatting gate is clean.
