+++ docs/features/active/2026-06-08-hierarchical-lcppn-folder-prediction-177/evidence/qa-gates/p1-nullable.2026-06-12T16-45.md
@@ -1,0 +1,16 @@
+# Phase 1 Nullable / Type-Check Build (Cycle 2)
+
+Timestamp: 2026-06-12T17:06Z
+
+Command: msbuild TaskMaster.sln /t:Build /p:Configuration=Debug /p:Platform="Any CPU" /p:Nullable=enable /p:TreatWarningsAsErrors=true
+(VS18 Community MSBuild. Incremental nullable gate convention.)
+
+EXIT_CODE: 0
+
+Output Summary:
+Build succeeded. 0 Warning(s), 0 Error(s). No new nullable warnings introduced by the two
+split files (LcppnFolderPredictor_Tests.cs / LcppnFolderPredictor_Classify_Tests.cs);
+zero diagnostics reference either file. Under /p:Nullable=enable the CS8632 annotations
+seen in the analyzer-build pass do not fire (nullable context is enabled solution-wide),
+so the protected nullable gate is clean. Pre-existing unrelated CS8625 in other files
+remain out of scope and were not modified.
